// cgocore-ui/src/app/features/models/reference/commodities/views/full/UnitOfMeasureFullView.ts
import { UnitOfMeasureForm } from 'src/app/features/models/reference/commodities/forms/UnitOfMeasureForm';
/**
 * auto generated full veiw bean ts file
 * <br/>write modifications between specific code marks
 * <br/>processed by skeleton-generator
 */
export class UnitOfMeasureFullView {

id: number;
form: UnitOfMeasureForm;
canUpdate: boolean;
canDelete: boolean;

/* Specific Code Start */
/* Specific Code End */
}